refactor(backend): update UsersModule
Conform to standard format from nest g resource
