+++ backend/src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { User } from './users/user.entity';
+import { User } from './users/entities/user.entity';
 import { UsersModule } from './users/users.module';
 import { MessagesModule } from './messages/messages.module';
 import { Message } from './messages/message.entity';
